requirements.txt
# ======================
# Deep Learning Framework
# ======================
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
pytorch-lightning>=2.1.0
lightning>=2.1.0
timm>=0.9.12
torchinfo>=1.8.0
torchmetrics>=1.2.0

# ======================
# Medical Imaging & Preprocessing
# ======================
monai>=1.3.0
albumentations>=1.3.1
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
imageio>=2.31.0
scikit-image>=0.21.0

# ======================
# Configuration & Experiment Tracking
# ======================
hydra-core>=1.3.2
omegaconf>=2.3.0
wandb>=0.16.0
tensorboard>=2.14.0
tensorboardX>=2.6.0

# ======================
# Data Processing & Analysis
# ======================
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
h5py>=3.9.0
pyarrow>=14.0.0

# ======================
# Visualization & XAI
# ======================
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0
grad-cam>=1.4.8
captum>=0.6.0

# ======================
# Deployment & API
# ======================
onnx>=1.15.0
onnxruntime>=1.16.0
onnxruntime-gpu>=1.16.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
aiofiles>=23.2.0
httpx>=0.25.0
requests>=2.31.0

# ======================
# Jupyter & Development
# ======================
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0
notebook>=7.0.0
nbformat>=5.9.0
tqdm>=4.66.0

# ======================
# Environment & Config
# ======================
python-dotenv>=1.0.0
pyyaml>=6.0.1
toml>=0.10.2

# ======================
# Testing & Code Quality
# ======================
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.6.0

# ======================
# Ordinal Regression
# ======================
# spacecutter>=0.3.0  # Not compatible with Python 3.12, using custom implementation instead

# ======================
# Utilities
# ======================
colorama>=0.4.6
rich>=13.6.0
click>=8.1.0
typer>=0.9.0
joblib>=1.3.0
multiprocess>=0.70.0
psutil>=5.9.0

# ======================
# Kaggle API
# ======================
kaggle>=1.5.16
